Removed axios http library, added Avatar component, added support for HTTP request abortion, added FileReader and FileCropper components. Added DialogHolder for more accurate callback triggering on Modal.Dialog render.

+++ Php/DevTools/Entity/Attributes/FileAttribute.php
@@ -57,7 +57,7 @@ class FileAttribute extends Many2OneAttribute
         parent::setValue($value, $fromDb);
 
         // If new files is being assigned and there is an existing file - delete the existing file
-        if (!$fromDb && $currentValue->id != $this->getValue()->id) {
+        if (!$fromDb && $currentValue && $currentValue->id != $this->getValue()->id) {
             $currentValue->delete();
         }
